[CMDs] Add Telegam CLI support commands
These are a couple of useful commands to start a Telegram service in
background and easily send commands to your peers.
The backgroud service is based on the Telegram CLI:
   https://github.com/vysheng/tg
and exploits STDIN injection to send commands without having to type them
in the console.

This opens plenty of possibilities for a scriptable telegram interaction.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>