[CMDs] Add script to update linux kernel CScope DB and VIM index
This script automate the update of the CScope DB for the currently
checked out kernel branch.
Each time a new branch is checked out, the CScope DB is build into
  ~/BUILDs/$PROJECT_DIR/$PROJECT_TAG
where:
  $PROJECT_DIR is the root folder of the kernel source tree
  $PROJECT_TAG is the name of the currently checked out branch

If the "~/BUILDs" folder does not exisits, the kernel source tree is
used as target for the CScope DB being generated.

If the "cscope.vim" VIM plugin is in use, this plugins expects
all the CScope DBs being available in the the ~/.cscope.vim folder.
In this case, this scripts ensure to properly link the generated
CScope DBs for the current kernel tree to make them availabe from
within future VIM sessions.
exists

a subfolder of the ~/BUILDs folder. This subfolder is named as the
main kernel source folder and a subfolder

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>