[Bootstrap] VIM: Build CommandT precompiled component
A binary component is required by the CommandT plugin.
This path builds it after at the first installation of
the plugin.

NOTE: a special version of VIM is required, which provides support for
ruby language (i.e. +ruby). This feature is provided by the vim-nox package
which contains a version of vim compiled with support for scripting with
Lua, Perl, Python, Ruby, and Tcl but no GUI.
Since I don't use the UI... ;-)

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>