Fix lookup for dotfiles folder
This version of dotfiles does not create anymore a ".dotfiles" symlink
pointing to our own dotfiles folder. This seems more simple since the
main repository is usually cloned just in the home directory.

This patch fixes all the dotfiles relative paths to point to the proper
folder (which is not more hidden).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>