[MUTT] Simplify configuration and installation
This patch update the default MUTT configuration to be more easy to
read and update by direct editing. The bootstrap script has been update
as well to just open the configuration file for editing.

It's up to the user to update the configuration file to select the
desired input/output method as well as the account data.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>