[CMDs] Add initial command to send Telegram messages from BASH
This patch adds a simple script which allows to easily send messages and
documents from BASH. The script requires a local installation of the
Telegram CLI (telegram-cli) which should be already configured to start
and login using a valid account. The script exploits the socket interface
exposed by the most recent version of telegram-cli.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>