[CMDs] Add support to send console output as HTML document
The existing switch "-c" has been extended to accept an option to specify
the kind of formatting for the console formatting.

Specifically "-c html" requires to send the console output a and HTML
document. NOTE: if you want to preserve the colors of the console input
remember to add "unbuffer" in front of the command generating the output.
For example
   unbuffer ls -la --color | tg_send -t UserName -c html

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>