[VIM] Add TaskPaper module
TaskPaper is a simple to-do list application that helps you stay
organized. Unlike competing applications, TaskPaper is based on
plain text files which offer you paper-like simplicity and ease
of use.

This patch adds to my VIM configuration the support for TaskPaper
file format as well as a couple of shortcuts which allows to
mark some tasks as urgent or pending.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>