[IMAPFilter] Add support for notification-only mode
The notification support provide by imapfilter is based on IMAP IDLE, thus
is a convenient and effective support to trigger actions on new mails being
available in the server, e.g. synchronize the local folder and/or send a
notification to the user.

This patch adds a configuration option which enable to run an imapfilter
instance without actually filtering mails while still generating notifications.
This support is useful in case you have a main instance of imapfilter always
running on a machine while on a secondary machine you still want to get
notified every time new mails are available.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>