[VIM] Reload configuration at each update
Every time we update the vim configuration file we need to reload the
editor or source the configuration to have the modifications in effect.

This patch simply monitor each write on disk of a buffer which name matches
that used by the vim configuration file and re-source it in the current
vim instance.

NOTE: this patch works only for console mode vim configuration files and
it assumes the file is in the standard location, i.e. ~/.vimrc

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>