[MUTT] Add a set of macros for manage patches on ML
This is a set of key bindings which allows to easily save and/or apply
patches embedded into email. For example, by tagging all the mails of a
patchset we can easily save the patches in mbox format by pressing (+t
while in index mode. The (+a key combination allows to apply the same
selected patches on top of the branch in the current working directory.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>