[MUTT] Add generic support for multiple addresses providers
This binds the query_command to a more generic set of scripts which
allows to recover email addresses from multiple sources, e.g. a local
alias DB and a remote LDAP server.

Currently, only local alias DBs are supported, which however:
- can be generated from multiple sources (e.g. offline LDAP queries)
- can be queried using a more generic expression

The main advantage of this method compare to the default mutt alias
menu, is that we can match any substring of an alias file, thus being
more flexible in addresses filtering and selection.

A query is triggered, by deafult, by pressing CTRL+T in an address query
field.

Inspired by:
- https://gitlab.com/muttmua/mutt/wikis/MuttGuide/Aliases
- https://gitlab.com/muttmua/mutt/wikis/QueryCommand

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>